Added methods to get state data
The trainer is now capable of getting the data regarding the state of the world.

The added methods allow getting whether the adjacent positions are walkable or not, whether the target is a certain direction or not, and a discretized value for the manhattan distance between the target and the agent.