Reworked action system
The trainer now has 2 methods to get actions.

GetRandomAction returns a random action.

GetBestAction returns the best action for a given state.

GetAction now retruns either a random direction or the best direction for a given state, based on the value of epsilon and a randomly generated value.

MoveAgent now makes use of GetActionDirection directly, since the action used is obtained through GetAction, which already accounts for the random value and epsilon exploration value.