Simulation stops when obtained reward is negative
When the agent performs an invalid action for which it gets punished, the simulation will stop the agen's Q values and then skip to the next simulation episode.

TODO: Fix problem where the agent can walk over the same cell as the follower, completely ignoring it's existance and overcoming the penalty of getting caught.