Added more testing scenes
These testing scenes are courtesy of Luangoma (https://github.com/Luangoma).

He passed me these maps so that I could test if the trained agent was capable of beating tough situations, specifically map number 4, which has a lot of 1 block wide tunnels that are impossible to escape from if the AI enters them.